Extra status: add support for extra status patch
The extra status patch in dwm adds support for splitting status text
between two bars. The position at which to split the status text is
indicated by a status separator (usually ';').

This commit adds support for copying the status separator to the status
text at the correct position and prevents an extra delimiter being
printed.

The status separator character and the index of the block at which the
split should occur are defined in config.h